+++ packages/server/src/database.ts
@@ -5,6 +5,8 @@ import { config } from '~/config';
 import {
   Customer,
   customerSchema,
+  Payment,
+  paymentSchema,
   Subscription,
   SubscriptionChange,
   subscriptionChangeSchema,
@@ -23,7 +25,7 @@ export class Database {
     this.orm = await MikroORM.init<PostgreSqlDriver>({
       type: 'postgresql',
       clientUrl: config.postgresUrl,
-      entities: [customerSchema, subscriptionSchema, subscriptionChangeSchema],
+      entities: [customerSchema, subscriptionSchema, subscriptionChangeSchema, paymentSchema],
       discovery: { disableDynamicFileAccess: true },
     });
   }
@@ -51,6 +53,10 @@ export class Database {
 
   get subscriptionChanges(): EntityRepository<SubscriptionChange> {
     return this.em.getRepository(SubscriptionChange);
+  }
+
+  get payments(): EntityRepository<Payment> {
+    return this.em.getRepository(Payment);
   }
 }
