+++ packages/server/src/entities/index.ts
@@ -1,5 +1,15 @@
 import { Customer, customerSchema } from './customer';
+import { Payment, paymentSchema } from './payment';
 import { Subscription, subscriptionSchema } from './subscription';
 import { SubscriptionChange, subscriptionChangeSchema } from './subscription_change';
 
-export { Customer, customerSchema, Subscription, SubscriptionChange, subscriptionChangeSchema, subscriptionSchema };
+export {
+  Customer,
+  customerSchema,
+  Payment,
+  paymentSchema,
+  Subscription,
+  SubscriptionChange,
+  subscriptionChangeSchema,
+  subscriptionSchema,
+};
