+++ packages/server/src/utils.ts
@@ -12,3 +12,12 @@ export function getPeriodFromAnchorDate(someDateInPeriod: Date, anchorDate: Date
     end: dayjs(someDateInPeriod).add(1, 'month').set('date', anchorDay).toDate(),
   };
 }
+
+export function getNextPeriodFromDate(date: Date, anchorDate: Date): { start: Date; end: Date } {
+  const anchorDay = Math.min(dayjs(date).add(1, 'month').endOf('month').date(), dayjs(anchorDate).date());
+  const nextDate = dayjs(date).add(1, 'month').set('date', anchorDay).toDate();
+  return {
+    start: dayjs(nextDate).set('date', anchorDay).toDate(),
+    end: dayjs(nextDate).add(1, 'month').set('date', anchorDay).toDate(),
+  };
+}
